Simplify v_out to uint8_t with proper voltage scaling
- get_v_out() now scales ADC reading to actual voltage (uint8_t, 0-255 = 0-255V)
- v_out register changed from union reg_u16 to uint8_t throughout
- Update protocol I2C write handler to match
- Minor refactoring in main loop for clarity
- Update VS Code MPLAB ioview settings

+++ mikona-firmware.X/mikona.h
@@ -17,7 +17,8 @@ enum led_color_t
 void set_led_color(enum led_color_t color);
 
 bool is_done(void);
-uint16_t get_v_out(void);
+// Returns voltage in volts, scaled to fit in uint8_t (0-255 = 0-4.096V)
+uint8_t get_v_out(void);
 
 void charge(bool charge);
 void discharge(bool enable);
